+++ fheroes2.vcxproj
@@ -189,7 +189,6 @@
     <ClCompile Include="src\engine\sdlnet.cpp" />
     <ClCompile Include="src\engine\serialize.cpp" />
     <ClCompile Include="src\engine\smk_decoder.cpp" />
-    <ClCompile Include="src\engine\sprites.cpp" />
     <ClCompile Include="src\engine\surface.cpp" />
     <ClCompile Include="src\engine\system.cpp" />
     <ClCompile Include="src\engine\thread.cpp" />
@@ -397,7 +396,6 @@
     <ClInclude Include="src\engine\sdlnet.h" />
     <ClInclude Include="src\engine\serialize.h" />
     <ClInclude Include="src\engine\smk_decoder.h" />
-    <ClInclude Include="src\engine\sprites.h" />
     <ClInclude Include="src\engine\surface.h" />
     <ClInclude Include="src\engine\system.h" />
     <ClInclude Include="src\engine\thread.h" />
