--- src/engine/sprites.cpp
@@ -1,62 +1,0 @@
-/***************************************************************************
- *   Copyright (C) 2012 by Andrey Afletdinov <fheroes2@gmail.com>          *
- *                                                                         *
- *   Part of the Free Heroes2 Engine:                                      *
- *   http://sourceforge.net/projects/fheroes2                              *
- *                                                                         *
- *   This program is free software; you can redistribute it and/or modify  *
- *   it under the terms of the GNU General Public License as published by  *
- *   the Free Software Foundation; either version 2 of the License, or     *
- *   (at your option) any later version.                                   *
- *                                                                         *
- *   This program is distributed in the hope that it will be useful,       *
- *   but WITHOUT ANY WARRANTY; without even the implied warranty of        *
- *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the         *
- *   GNU General Public License for more details.                          *
- *                                                                         *
- *   You should have received a copy of the GNU General Public License     *
- *   along with this program; if not, write to the                         *
- *   Free Software Foundation, Inc.,                                       *
- *   59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.             *
- ***************************************************************************/
-
-#include "sprites.h"
-#include "display.h"
-
-SpritePos::SpritePos() {}
-
-SpritePos::SpritePos( const Surface & sf, const Point & pt )
-    : Surface( sf )
-    , pos( pt )
-{}
-
-const Point & SpritePos::GetPos( void ) const
-{
-    return pos;
-}
-
-Rect SpritePos::GetArea( void ) const
-{
-    return Rect( GetPos(), GetSize() );
-}
-
-void SpritePos::SetSurface( const Surface & sf )
-{
-    Surface::Set( sf, true );
-}
-
-void SpritePos::SetPos( const Point & pt )
-{
-    pos = pt;
-}
-
-void SpritePos::Reset( void )
-{
-    pos = Point( 0, 0 );
-    Surface::Reset();
-}
-
-u32 SpritePos::GetMemoryUsage( void ) const
-{
-    return Surface::GetMemoryUsage() + sizeof( pos );
-}
